clear sudo context
Even tho the next documented command is `cd /opt`, changing to root with `sudo su` will keep the current user context, like the current user dir (and other stuffs). By using `-` it will clear the current user context, and the current dir will be the root's home dir. This is a cleaner approach when changing to root.