theme: set <base> to the correct path from RTD
<base> path should not be hardcoded during the release, because the
exact same code can be built into different paths by Read The Docs.

For example, the last release 2.20.0 is built under "/en/2.20.0/", but
it also gets built under "/en/stable/".

It is possible to query RTD about the correct value for the path, so do
that in order to provide a correct <base> and working URLs in all docs.