theme: reword version warning
Clearer phrase, now that we're paving the road to have "/stable/" and
"/latest/" URLs, where stable is the latest stable version that has been
released, and latest is the last development commit (typicaly from the
"master" or "main" branch)

Inspired by the banners that appear in the Jinja documentation:

latest: https://jinja.palletsprojects.com/en/latest/
Warning: This is the development version. The latest stable version is
Version 3.0.x.

old: https://jinja.palletsprojects.com/en/2.11.x/
Warning: This is an old version. The latest stable version is Version
3.0.x.